+++ cli/src/tests/mod.rs
@@ -1,1 +1,2 @@
 mod output_tests;
+mod shell_tests;