+++ node/src/tests/mod.rs
@@ -1,5 +1,0 @@
-mod local_node_tests;
-mod support;
-
-#[cfg(feature = "http")]
-mod http_api_tests;